Create a separate go module for pinniped-auth plugin
