+++ cmd/cli/plugin/pinniped-auth/login_test.go
@@ -18,7 +18,7 @@ import (
 	"github.com/MakeNowJust/heredoc"
 	"github.com/stretchr/testify/require"
 
-	"github.com/vmware-tanzu/tanzu-framework/pkg/v1/buildinfo"
+	"github.com/vmware-tanzu/tanzu-framework/cli/runtime/buildinfo"
 )
 
 const (
